+++ src/primes.rs
@@ -12,7 +12,7 @@ use crate::bigint;
 /// of greater than 32 could lead to public keys that were unable to be sent over the network.
 ///
 #[doc(alias = "N")]
-pub const LARGE_SAFE_PRIME_LENGTH: usize = 32;
+pub const LARGE_SAFE_PRIME_LENGTH: u8 = 32;
 
 /// Static large safe prime (`N`) value.
 /// The big endian version of [LARGE_SAFE_PRIME_LITTLE_ENDIAN].
@@ -30,7 +30,7 @@ pub const LARGE_SAFE_PRIME_LENGTH: usize = 32;
 ///  hex!("894B645E89E1535BBDAD5B8B290650530801B18EBFBF5E8FAB3C82872A3E9BB7"));
 /// ```
 #[doc(alias = "N")]
-pub const LARGE_SAFE_PRIME_BIG_ENDIAN: [u8; LARGE_SAFE_PRIME_LENGTH] = [
+pub const LARGE_SAFE_PRIME_BIG_ENDIAN: [u8; LARGE_SAFE_PRIME_LENGTH as usize] = [
     137_u8, 75_u8, 100_u8, 94_u8, 137_u8, 225_u8, 83_u8, 91_u8, 189_u8, 173_u8, 91_u8, 139_u8,
     41_u8, 6_u8, 80_u8, 83_u8, 8_u8, 1_u8, 177_u8, 142_u8, 191_u8, 191_u8, 94_u8, 143_u8, 171_u8,
     60_u8, 130_u8, 135_u8, 42_u8, 62_u8, 155_u8, 183_u8,
@@ -52,14 +52,14 @@ pub const LARGE_SAFE_PRIME_BIG_ENDIAN: [u8; LARGE_SAFE_PRIME_LENGTH] = [
 ///  hex!("b79b3e2a87823cab8f5ebfbf8eb10108535006298b5badbd5b53e1895e644b89"));
 /// ```
 #[doc(alias = "N")]
-pub const LARGE_SAFE_PRIME_LITTLE_ENDIAN: [u8; LARGE_SAFE_PRIME_LENGTH] = [
+pub const LARGE_SAFE_PRIME_LITTLE_ENDIAN: [u8; LARGE_SAFE_PRIME_LENGTH as usize] = [
     183_u8, 155_u8, 62_u8, 42_u8, 135_u8, 130_u8, 60_u8, 171_u8, 143_u8, 94_u8, 191_u8, 191_u8,
     142_u8, 177_u8, 1_u8, 8_u8, 83_u8, 80_u8, 6_u8, 41_u8, 139_u8, 91_u8, 173_u8, 189_u8, 91_u8,
     83_u8, 225_u8, 137_u8, 94_u8, 100_u8, 75_u8, 137_u8,
 ];
 
 pub(crate) struct LargeSafePrime {
-    prime: [u8; LARGE_SAFE_PRIME_LENGTH],
+    prime: [u8; LARGE_SAFE_PRIME_LENGTH as usize],
 }
 
 impl Default for LargeSafePrime {
@@ -70,11 +70,11 @@ impl Default for LargeSafePrime {
     }
 }
 impl LargeSafePrime {
-    pub const fn from_le_bytes(prime: &[u8; LARGE_SAFE_PRIME_LENGTH]) -> Self {
+    pub const fn from_le_bytes(prime: &[u8; LARGE_SAFE_PRIME_LENGTH as usize]) -> Self {
         Self { prime: *prime }
     }
 
-    pub const fn as_le_bytes(&self) -> &[u8; LARGE_SAFE_PRIME_LENGTH] {
+    pub const fn as_le_bytes(&self) -> &[u8; LARGE_SAFE_PRIME_LENGTH as usize] {
         &self.prime
     }
